# bot/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "dYdX Trading Bot API",
    "description": "API for managing multiple dYdX trading bot instances with JWT Authentication. Most non-auth HTTP endpoints return a standardized envelope: {success, message, data, timestamp, trace_id}.",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/strategies": {
      "get": {
        "tags": [
          "Strategies"
        ],
        "summary": "List Strategies",
        "description": "List stored strategies for the UI.",
        "operationId": "list_strategies_api_v1_strategies_get",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "skip",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 0,
              "title": "Skip"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 50,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/StandardApiResponse"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {}
                      }
                    }
                  ]
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "x-response-envelope": "StandardApiResponse"
      },
      "post": {
        "tags": [
          "Strategies"
        ],
        "summary": "Create Strategy",
        "description": "Create a strategy.",
        "operationId": "create_strategy_api_v1_strategies_post",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StrategyRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/StandardApiResponse"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {}
                      }
                    }
                  ]
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "x-response-envelope": "StandardApiResponse"
      }
    },
...
    "/api/v1/strategies/{strategy_id}": {
      "get": {
        "tags": [
          "Strategies"
        ],
        "summary": "Get Strategy",
        "description": "Get one strategy.",
        "operationId": "get_strategy_api_v1_strategies__strategy_id__get",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "strategy_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Strategy Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/StandardApiResponse"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {}
                      }
                    }
                  ]
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "x-response-envelope": "StandardApiResponse"
      },
      "put": {
        "tags": [
          "Strategies"
        ],
        "summary": "Update Strategy",
        "description": "Update one strategy.",
        "operationId": "update_strategy_api_v1_strategies__strategy_id__put",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "strategy_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Strategy Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StrategyRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/StandardApiResponse"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {}
                      }
                    }
                  ]
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "x-response-envelope": "StandardApiResponse"
      },
      "delete": {
        "tags": [
          "Strategies"
        ],
        "summary": "Delete Strategy",
        "description": "Delete one strategy.",
        "operationId": "delete_strategy_api_v1_strategies__strategy_id__delete",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "strategy_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Strategy Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/StandardApiResponse"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {}
                      }
                    }
                  ]
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "x-response-envelope": "StandardApiResponse"
      }
    },
    "/api/v1/strategies/{strategy_id}/versions": {
      "get": {
        "tags": [
          "Strategies"
        ],
        "summary": "Get Strategy Versions",
        "description": "Get in-memory version history for a strategy.",
        "operationId": "get_strategy_versions_api_v1_strategies__strategy_id__versions_get",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "strategy_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Strategy Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/StandardApiResponse"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {}
                      }
                    }
                  ]
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "x-response-envelope": "StandardApiResponse"
      }
    },
    "/api/v1/strategies/{strategy_id}/versions/{version_id}/revert": {
      "post": {
        "tags": [
          "Strategies"
        ],
        "summary": "Revert Strategy Version",
        "description": "Revert a strategy to a prior stored version.",
        "operationId": "revert_strategy_version_api_v1_strategies__strategy_id__versions__version_id__revert_post",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "strategy_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Strategy Id"
            }
          },
          {
            "name": "version_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Version Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StrategyVersionRevertRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/StandardApiResponse"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {}
                      }
                    }
                  ]
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "x-response-envelope": "StandardApiResponse"
      }
    },
...
    "/api/v1/backtests": {
      "post": {
        "summary": "Create Backtest",
        "description": "Create and start a new backtest",
        "operationId": "create_backtest_api_v1_backtests_post",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/BacktestConfigRequest"
                  },
                  {
                    "$ref": "#/components/schemas/BacktestRunRequestCompat"
                  }
                ],
                "title": "Request"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/StandardApiResponse"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/BacktestResponse"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "x-response-envelope": "StandardApiResponse"
      },
      "get": {
        "summary": "List Backtests",
        "description": "List backtest runs with filtering",
        "operationId": "list_backtests_api_v1_backtests_get",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 0,
              "title": "Offset"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "name": "days",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Days"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/StandardApiResponse"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/BacktestListResponse"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "x-response-envelope": "StandardApiResponse"
      }
    },
...
    "/api/v1/backtests/{run_id}": {
      "get": {
        "summary": "Get Backtest Details",
        "description": "Get detailed backtest results",
        "operationId": "get_backtest_details_api_v1_backtests__run_id__get",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/StandardApiResponse"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/BacktestDetailResponse"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "x-response-envelope": "StandardApiResponse"
      },
      "delete": {
        "summary": "Delete Backtest",
        "description": "Delete backtest run and all associated data",
        "operationId": "delete_backtest_api_v1_backtests__run_id__delete",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/StandardApiResponse"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {}
                      }
                    }
                  ]
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "x-response-envelope": "StandardApiResponse"
      }
    },
...
    "/api/v1/backtests/{run_id}/create-strategy": {
      "post": {
        "summary": "Create Strategy From Backtest",
        "description": "Create a strategy snapshot from an existing backtest.",
        "operationId": "create_strategy_from_backtest_api_v1_backtests__run_id__create_strategy_post",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BacktestCreateStrategyRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/StandardApiResponse"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {}
                      }
                    }
                  ]
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "x-response-envelope": "StandardApiResponse"
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BacktestCreateStrategyRequest": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "Name"
          },
          "description": {
            "type": "string",
            "maxLength": 2000,
            "title": "Description",
            "default": ""
          },
          "config": {
            "additionalProperties": true,
            "type": "object",
            "title": "Config"
          }
        },
        "type": "object",
        "required": [
          "name"
        ],
        "title": "BacktestCreateStrategyRequest",
        "description": "Create a strategy from an existing backtest."
      },
...
      "BacktestListResponse": {
        "properties": {
          "backtests": {
            "items": {
              "$ref": "#/components/schemas/BacktestResponse"
            },
            "type": "array",
            "title": "Backtests"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          }
        },
        "type": "object",
        "required": [
          "backtests",
          "total"
        ],
        "title": "BacktestListResponse",
        "description": "Backtest list response"
      },
...
      "BacktestResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "status": {
            "$ref": "#/components/schemas/BacktestStatus"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          },
          "config": {
            "$ref": "#/components/schemas/BacktestConfigRequest"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "status",
          "created_at",
          "updated_at",
          "config"
        ],
        "title": "BacktestResponse",
        "description": "Backtest response"
      },
...
      "BacktestStatus": {
        "type": "string",
        "enum": [
          "pending",
          "running",
          "paused",
          "completed",
          "failed",
          "timeout",
          "cancelled",
          "stale"
        ],
        "title": "BacktestStatus",
        "description": "Backtest status enumeration"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "StrategyVersionRevertRequest": {
        "properties": {},
        "type": "object",
        "title": "StrategyVersionRevertRequest",
        "description": "Placeholder body for strategy version revert."
      },
...
      "StandardApiResponse": {
        "type": "object",
        "required": [
          "success",
          "message",
          "data",
          "timestamp",
          "trace_id"
        ],
        "properties": {
          "success": {
            "type": "boolean"
          },
          "message": {
            "type": "string"
          },
          "data": {
            "type": [
              "object",
              "array",
              "string",
              "number",
              "boolean",
              "null"
            ]
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          },
          "trace_id": {
            "type": "string"
          }
        },
        "description": "Standard API response envelope used by bot/runtime HTTP endpoints."
      }
    },
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Enter your JWT token"
      }
    }
  }
}